Add files for containerized NFS server
This commit adds the beginning of a build for a containerized NFS
server. This build is based on Debian 12 and uses SystemD to
run the various NFS components.

Only v4 of the NFS protocol is enabled.

The container must be run in privileged mode so that the nfsd
and pipefs filesystems can be mounted.