Downgrade Playwright to v1.53.2
The 1.54.0 release of Playwright added a `denyLaunch` parameter that
basically causes the browser server to respond to every request
with:

    Launching more browsers is not allowed.

It seems that this is kinda, sorta, maybe a guardrail for the future
implementation of a MCP server that can be manipulated by AI models.

However, there is no commit description that explains the rationale
behind this, and no documented way to disable it so that the
server is actually usable.

So, this commit pins the containerized version of Playwright back to
the last release before the change: 1.53.2. This version appears
to work well with the 1.54.0 release of playwright-ruby-client.