Update playwright to 1.55.1
This commit updates the playwright build to the latest
1.55.1 version.

NOTE: There was a breaking API change in v1.54.0 that
made it so `Playwright.connect_to_playwright_server`
could not launch new browser instances. This
is fixed by a combination of the 1.54.1 release of
playwright-ruby-client and switching to the
`Playwright.connect_to_browser_server` method.

However, this method does not expose browser
options. So, launching a "headful" browser
requires patching in the `x-playwright-launch-options`
header.