Migrate Headscale to install via UBI
This commit re-factors the Headscale container to install the binary
using the Universal Binary Installer instead of a curl command.

VOLUMES and PORTS are also updated/added to make the config work
better with Docker Compose and Traefik ingress.

The CMD is also changed to ENTRYPOINT to simplify command execution
within the container: bootstrapping user accounts, generating API
keys, etc.