Add Mise tooling to project
This commit adds a `mise.toml` configuration file that sets up the
`mise` configuration tool to drive build and test tasks:

  - Installation of Ruby Gems
  - Building docker container, with re-builds conditional on changes to
    input files
  - Staging acceptance tests in Docker Compose
  - Running Docker acceptance tests

Specific build and test flows are implemented for the `playwright`
container.

Most of this logic was extracted from the `build_container.yaml`
Github Action. Eventually the configuration will be extended to
other containers, to Kubernetes tests, and finally the Github
Action will be updated to run Mise tasks.